fix: use precise finder api cooldown cutoffs
Finder already preserved fractional cooldown values and rounded successful checkpoints up, but it still floored the current-time snapshot before evaluating source eligibility.

Using the precise wall-clock time keeps cooldown expiry aligned with the configured threshold instead of delaying eligible API sources by nearly a second.

+++ src/bigbrotr/services/finder/api_runtime.py
@@ -40,7 +40,7 @@ class ApiDiscoveryWorkerContext:
 
     brotr: Brotr
     cooldown: float
-    now: int
+    now: float
     max_response_size: int
     request_delay: float
     is_running: Callable[[], bool]
@@ -68,7 +68,7 @@ def build_api_source_attempts(
     checkpoint_map: Mapping[str, ApiCheckpoint],
     *,
     cooldown: float,
-    now: int,
+    now: float,
     logger: Logger,
 ) -> tuple[ApiSourceAttempt, ...]:
     """Return the enabled API sources whose cooldown has elapsed for this cycle."""
@@ -91,7 +91,7 @@ async def stream_api_discovery_attempts(  # noqa: PLR0913
     checkpoint_map: Mapping[str, ApiCheckpoint],
     *,
     cooldown: float,
-    now: int,
+    now: float,
     max_response_size: int,
     request_delay: float,
     is_running: Callable[[], bool],
